requirements.txt
Flask==3.0.2
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
pkg_resources==0.0.0
python-dotenv==1.0.1
python-pptx==0.6.23
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.1
XlsxWriter==3.2.0
zipp==3.17.0
